admin-fastapi/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# ORM + Database
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29.0
alembic>=1.13.0

# Redis
redis[hiredis]>=5.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Captcha
easy-captcha-python>=1.0.0

# Object Storage
minio>=7.2.0

# System Monitoring
psutil>=5.9.0

# Configuration
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Utilities
python-multipart>=0.0.9
httpx>=0.27.0
